feat(cypress-tags.js): scenario outlines with examples were not correctly filtered by tags (#609)
* fix(cypress-tags.js): scenario outlines with examples were not correctly filtered by tags

Now the script supports filtering examples of scenario outlines too. I have removed the optimization
that we should check scenario tags if there are any, because they do not work for negated tags (e.g.
`not @tag1`). Additionally, just checking the feature tag itself is also wrong, because a feature
with tag `@feature` having a scenario with tag `@scenario` would incorrectly run when we filter for
`not @scenario`. So the best is to leave the tag checking for "leaf nodes", i.e. scenarios or
examples. Also consider the corner case where a feature has no scenarios or scenario outlines at
all, i.e. again we would not run such a feature at all.

Fixes #196.

* fixup! fix(cypress-tags.js): scenario outlines with examples were not correctly filtered by tags